Move zsh under .config
Zsh doesn't support XDG conventions and therefore this needs all sorts
of hacks, and `.zshenv` still stays under `~`, as `ZDOTDIR` is defined
there. Despite these, the advantages are that zsh will not litter `~`
with history and dump files, and `.zshrc` can be tinkered with even on
Windows without running the bootstrap script over and over again.

However, the bootstrap script is still broken on Windows.