check_msys_processes: Only check installed optional dependencies
This change is to avoid false positives. Imagine we have leaf
packages, 'p1' and 'p2'. 'p1' has an optional dependency 'd1' with a
hard dependency on 'd2', and 'p2' with a hard dependency on 'd2'. 'd1'
is not installed. Without this change if a program from 'p1' is
running, and 'd2' is outdated, we would prevent upgrades.

If a package is an optional dependency and has further dependencies but is